# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Simpul HR API
  version: 1.0.0
  description: Multi-tenant REST API for the Simpul HR management platform.
servers:
  - url: http://localhost:8080/v1
security:
  - bearerAuth: []
...
paths:
  /auth/login:
    post:
      tags: [Auth]
      security: []
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200': { description: Session or organization selection, content: { application/json: { schema: { $ref: '#/components/schemas/SessionResponse' } } } }
        '202': { description: MFA challenge, content: { application/json: { schema: { $ref: '#/components/schemas/MFAChallenge' } } } }
        default: { $ref: '#/components/responses/Problem' }
  /auth/refresh:
    post:
      tags: [Auth]
      security: []
      operationId: refreshSession
      responses:
        '200': { description: Rotated session, content: { application/json: { schema: { $ref: '#/components/schemas/SessionResponse' } } } }
        default: { $ref: '#/components/responses/Problem' }
...
  /auth/mfa/enroll:
    post:
      tags: [Auth]
      security: []
      operationId: enrollMfa
      requestBody: { $ref: '#/components/requestBodies/MFAChallengeToken' }
      responses:
        '200': { description: TOTP enrollment, content: { application/json: { schema: { type: object, required: [secret, otpauthUri], properties: { secret: { type: string }, otpauthUri: { type: string, format: uri } } } } } }
        default: { $ref: '#/components/responses/Problem' }
  /auth/mfa/confirm:
    post:
      tags: [Auth]
      security: []
      operationId: confirmMfa
      requestBody: { $ref: '#/components/requestBodies/MFACode' }
      responses:
        '200': { description: Session and recovery codes, content: { application/json: { schema: { $ref: '#/components/schemas/SessionResponse' } } } }
        default: { $ref: '#/components/responses/Problem' }
  /auth/mfa/verify:
    post:
      tags: [Auth]
      security: []
      operationId: verifyMfa
      requestBody: { $ref: '#/components/requestBodies/MFACode' }
      responses:
        '200': { description: Session, content: { application/json: { schema: { $ref: '#/components/schemas/SessionResponse' } } } }
        default: { $ref: '#/components/responses/Problem' }
  /auth/sso/{provider}/start:
    get:
      tags: [Auth]
      security: []
      operationId: startSso
      parameters: [ { $ref: '#/components/parameters/Provider' }, { name: organizationSlug, in: query, required: false, schema: { type: string } } ]
      responses:
        '200': { description: Provider authorization URL, content: { application/json: { schema: { type: object, properties: { authorizationUrl: { type: string } } } } } }
        default: { $ref: '#/components/responses/Problem' }
  /auth/sso/{provider}/callback:
    get:
      tags: [Auth]
      security: []
      operationId: finishSso
      parameters: [ { $ref: '#/components/parameters/Provider' }, { name: code, in: query, schema: { type: string } }, { name: state, in: query, schema: { type: string } } ]
      responses: { '302': { description: Redirect to SPA }, default: { $ref: '#/components/responses/Problem' } }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  parameters:
...
    Provider: { name: provider, in: path, required: true, schema: { type: string, enum: [google, entra, slack] } }
...
  requestBodies:
    MFAChallengeToken: { required: true, content: { application/json: { schema: { type: object, required: [challengeToken], properties: { challengeToken: { type: string } } } } } }
    MFACode: { required: true, content: { application/json: { schema: { type: object, required: [challengeToken, code], properties: { challengeToken: { type: string }, code: { type: string } } } } } }
...
  responses:
    Problem: { description: Problem details, content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } } }
...
  schemas:
    Problem:
      type: object
      required: [type, title, status, code, detail]
      properties: { type: { type: string, format: uri }, title: { type: string }, status: { type: integer }, code: { type: string }, detail: { type: string }, errors: { type: object, additionalProperties: { type: string } }, requestId: { type: string } }
...
    LoginRequest: { type: object, required: [email, password], properties: { email: { type: string, format: email }, password: { type: string, minLength: 8 }, organizationSlug: { type: string } } }
    MFAChallenge: { type: object, properties: { mfaRequired: { type: boolean }, mfaEnrollmentRequired: { type: boolean }, challengeToken: { type: string } } }
    SessionResponse: { type: object, properties: { accessToken: { type: string }, expiresIn: { type: integer }, user: { $ref: '#/components/schemas/SessionUser' }, recoveryCodes: { type: array, items: { type: string } }, organizationSelectionRequired: { type: boolean }, organizations: { type: array, items: { $ref: '#/components/schemas/WorkspaceOption' } } } }
    WorkspaceOption: { type: object, required: [slug, name], properties: { slug: { type: string }, name: { type: string } } }
    SessionUser: { type: object, properties: { id: { type: string, format: uuid }, email: { type: string, format: email }, role: { type: string }, employeeId: { type: string, format: uuid }, permissions: { type: array, items: { type: string } }, organization: { $ref: '#/components/schemas/Organization' }, mfaEnabled: { type: boolean } } }
    Organization: { allOf: [ { $ref: '#/components/schemas/OrganizationInput' }, { type: object, properties: { id: { type: string, format: uuid }, slug: { type: string }, createdAt: { type: string, format: date-time }, updatedAt: { type: string, format: date-time } } } ] }
    OrganizationInput: { type: object, properties: { name: { type: string }, industry: { type: string }, address: { type: string }, phone: { type: string }, email: { type: string, format: email }, timezone: { type: string }, locale: { type: string }, currency: { type: string } } }
...